Fix Cross-compilation.rst header hierarchy
Cross-compilation.rst was using the page header format for sections headers. This caused the tutorial's list to display the sections instead of the full tutorial, what could be confusing. For example an 'Overview' tutorial would be expected to be a ROS2 overview, instead of a Cross-Compilation overview.
This change adds a page header and lowers the others headers one step in the hierarchy, based on Colcon-Tutorial.rst. It also adds a Table of Contents and the `redirect-from` info to keep it similar to the rest of the tutorials.