Update installation files for their new home and add redirects.
For previous ROS 2 releases we've stacked installation instructions on
top of each other and used prose to indicate when instructions pertain
to or exclude certain releases.

As we add more releases the prose gets more convoluted and harder for
readers to follow. Instead we'll create a bit of duplication in order to
keep the installation instructions for Crystal (and earlier) intact
while making edits for Dashing in a separate parallel directory
(forthcoming).

I have addressed sphinx build warnings locally and have changed several
instances of `../Tutorials/...`  to `/Tutorials/...` so that the
relative depth of tutorials and the installation docs isn't significant.
At least testing locally with `make html` these are rendered into html
as relative links so they work regardless of the directory structure
above the sphinx root.

---

Add redirects to unversioned installation docs.

Also restores the "generic" link in other pages that point to when they
want the "current" install docs. We'll update the default rosdistro by
moving the redirect target.

Resolve references to Installation by routing them to Installation/Crystal (#179)

* Resolve missing references. (#183)

Signed-off-by: Michel Hidalgo <michel@ekumenlabs.com>