Yield content for all redirects. (#178)
This was likely an indentation bug.
With this change content is yielded once per redirect_url rather than
only once per canonical_url. A side effect of the previous revision is
that if a source file redirects from multiple paths only one of those
redirects (specifically the last one to be iterated) will be placed.