Updates to the Fedora development setup.
Updates to support the build-from-source instructions.  Tested by
following the directions on the wiki in a fedora:30 docker container.

python3-lark-parser-0.7.1 is available in f29 and f30 - remove the pip
installation.

colcon needs to be installed via pip, and requires a version of pytest
greater than the version available in the Fedora repos (3.9).  --upgrade
is required to install over the system-wide 3.9 version.

wget and rosdep are required by the build from source instructions, and
not installed by default.

gcc-c++, patch, and make are required to build C++ packages, and may not
be installed on a fresh system. redhat-rpm-config provides hardening
flags that gcc-c++ needs.

liblsan is needed to build rosbag2 (the sanitizer isn't installed by
default when gcc-c++ is installed).  It should probably be disabled by
default, and/or added as a system rosdep.

Signed-off-by: Rich Mattes <richmattes@gmail.com>