reword install Prerequisites section
I found the section about getting the demo a little confusing, so I reworded it. First the text made it sound like there were two OpenCV install links, but one of them was actually a link to installing Crystal from source. Second it was missing that `image_tools` needs to be installed.