Specify that development installation instructions are shared with Rolling.
In discussion between Marya and myself we've gone back and forth on what
information about the Rolling distribution we want to be put on this
page. We think we want most people to find the Rolling instructions only
when following the development setup or prerelease testing workflows.
Thus, the Rolling installation pages are linked by the Latest
Development Setup via the $PLATFORM-latest anchors. The pre-release
binaries (which would also be covered by the Rolling installation
instructions) link back to the Latest Development Setup for their
installation instructions so these two will arrive at the Rolling
instructions.

However, without this change, if a reader comes to this page explicitly
seeking the installation instructions for the Rolling release they will
find no reference to them. No other documentation suggests or
recommends installing the rolling release except to support a prerelease
binary or latest development installation so we think this small
addition, which takes those looking to answer "How do I install the
rolling distribution" to an explaination of the rolling distribution's
role and which suggests that most individuals actually want the stable
distribution. Yet those who require the Rolling installation
instructions for Pre-release or development setup will still be guided
to the rolling installation instructions in those cases.  arbitrarily
seeking the Rolling installation instructions