Update Linux-Development-Setup.rst
If you have build another workspace (e.g. the turtlebot workspace) and sourced that one, you also source the previous ROS version. Someone in our group ran into this issue. You get a lot of build errors in that case. To make sure that there is no ROS sourced in the terminal, you can check it.