Cleanup github actions.
1.  Add in the sphinx-multiversion pip dependency.  It isn't
required on the master branch, but it doesn't hurt anything
and is needed on the rolling, foxy, and dashing branches.

2.  Github actions fail immediately if the command returns
non-zero.  Since we were capturing the output, this was
effectively hiding the error from us.  We now just check
things sequentially.  This is better in that it is simpler,
but worse in that you don't get all of the errors at once.
I still like the simpler approach.

3.  Get rid of the exclusions for the action tutorials, as
it doesn't seem to be needed anymore.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>