Use the html_baseurl configuration variable.
Unfortunately, in a world where we are using sphinx multiversion,
this isn't quite enough since we can't embed the current distribution
name into the config.  Also add in a hook that will rewrite the
html_baseurl right after the configuration has been finished
initializing.

While we are in here, also update a couple of URLs and dates.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>