Fix up the distribution table. (#1179)
In particular, make it so that the currently active distributions
are marked in green.  To do this, I unfortunately had to result
to raw HTML and CSS, since I could not figure out a way to make
RST play nice with the Read-the-docs theme to make this happen.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit d80685b6ccf566f8046ff7e2d013454e9ae40b9f)