Fix references to index.ros.org.
In most cases, make them relative so we don't have to go
outside to get the references.  In at least one case, just
remove the link completely.

With this commit in place, the only references to index.ros.org
have to do with package-specific documentation, and some links
in the release notes that I was reluctant to change.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>