Start a draft for an Ament Python guide.
This is very much a work in progress and is littered with implicit
TODOs in addition to lacking much organization.

I started by trying to jot down as many major questions or known
limitations of ament_python packages as I could think of.

I'm going to keep iterating on this as a draft but if anyone else would
like to throw on extra info I invite anyone on the team to toss a few
commits on the branch.