Put all of the information about releases in conf.py. (#1409)
Previously we had put some hard-coded lists of EOL distributions
in some page templates.  However, it is much better to have
all of the information about distributions in the configuration
file.  To support this, add in a new config variable 'smv_eol_versions'
that has a list of the EOL versions.  When in multiversion mode,
propagate this into the page context so that the pages can just
reference the variable.

Finally, while we are in here, take galactic out of the "released"
list until it is actually released.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>