Don't propose to simply close a terminal (#1643)
Terminals offer a command line into the system where you can start
arbitrary and multiple processes. Closes the terminal does not guarantee
that all subprocesses will be closed as well. Even more, there are
situations where you don't want to have the subprocesses closed, for
example when you have an editor with unsaved data suspended in the
background.

As this is an introduction tutorial probably read by a lot of people who
just start with the command line and given that Ctrl+C is easy to reach,
I propose to drop the comment to close terminals.