Fix for Windows build failure
Building ROS2 with the current guide fails due to a compilation error of zstd_vendor. See https://github.com/ros2/rosbag2/issues/630 and https://github.com/ros2/rosbag2/issues/834 for examples of this.

The reason is that there is no patch.exe in Windows. Fortunately Git installs one. It only needs to be linked through the PATH environment variable.