Update Foxy's conf.py so we can build standalone. (#1849)
One of the easier way to work on branches in this documentation
is to make local changes and then run "make html".  This works
fine on the rolling and galactic branches, but because Foxy's
conf.py was so out-of-date, it didn't really work here.  Update
the conf.py so it is compatible with the galactic one, which
makes it work a lot better.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>