Add a couple of extra steps for Windows Debug mode. (#1960)
While working in Windows debug mode, I noticed that it couldn't
import pytest, and hence none of the pytest tests in the core
were running.  After some debugging, I found that pytest (or
more specifically, pytest-cov/coverage) requires the sqlite3_d.dll
to be in place for it to work.  Add the instruction to copy that
over, as well as add a test to make sure it worked.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit 83a3b11b80fbbcc47f393145cdc5a776320d6f21)