Update Creating-A-Workspace.rst
The documentation says that the 3 first packages are ignored, but it is confusing because it does not say why they are ignored. I added text to specify that colcon ignores those packages because a IGNORE file is present. It would be even better to explain why this file is present, but I don't know, I am not familiar with these demos.