1.2 Publishing Feedback
In section 1.2 there is some code missing in the python script. After lines 141-142 in the instructions, the following lines of code should be included at the end of the script: 

def main(args=None):
    rclpy.init(args=args)

    fibonacci_action_server = FibonacciActionServer()

    rclpy.spin(fibonacci_action_server)

if __name__ == '__main__':
    main()

Currently, if users copy and paste the current python script, then try and initialize it in the terminal it won't work.