Update Using-ROS2-Launch-For-Large-Projects.rst
Just added some additional information the the prerequisites, as it is not clear to new comers that the package needs to be build for python in order to handle launch files. I build the package by cmake and looked all over for the setup.py, only to realize that I should have build the package for python in the first place.