Update to ROS launch tutorial  (#2084)
Just added some additional information the the prerequisites, as it is not clear to new comers that the package needs to be build for python in order to handle launch files. I build the package by cmake and looked all over for the setup.py, only to realize that I should have build the package for python in the first place.

Co-authored-by: Chris Lalancette <clalancette@gmail.com>
(cherry picked from commit 010b31eb352a3cfd33f721d4e473f683a7f860d3)