Remove Galactic-specific wording. (#2422)
There is no reason for this wording; we have separate documentation
pages for each of the released distributions.

Because of this, this commit should be backported to the
'galactic' branch, but not the 'foxy' one.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit eaf6e9bfa763dc8e0aa04f3c48adfdf5d31c2fcc)