Fix links to installation instructions. (#2797)
That is, since we've switched to one-distro-per-branch,
these links were broken.

This is kind of a silly fix since this is to a long EOL
distribution, but we may as well have all of the links
working.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit b7cbc0406d6c185598ea5b44fad4ce93ed1c6548)

# Conflicts:
#	source/Releases/Release-Dashing-Diademata.rst