Add a reference to the new source entry requirement for ros2-gbp repository creation.
As discussed, we're adding a requirement to the release repository
creation process to add a source stanza, ensuring that the packages in
the new repository undergo rosdistro review before a release repository
is created. This is particularly important for naming convention review
since that review could affect the name of the source, and by extension,
release repository created.