Update the code in the tf2 tutorials to correspond to the latest. (#3024)
This commit actually does a full audit of the Tf2 tutorials to do
a number of things in them:

1.  Synchronizes the code in here with what is in
    https://github.com/ros/geometry_tutorials.  In the process of doing
    this, it fixes some small bugs.
2.  It makes the tutorials far more internally consistent, so that we
    have the same set of steps for each tutorial and each step is
    well-defined.
3.  It splits the "Build and run" sections into separate "Build" and
    "Run" sections, as that just seemed to make more sense.
4.  It fixes up the spacing so that it is all internally consistent
    within the tutorials.  This makes up the bulk of the changes, but I
    think it is worthwhile to do.

One thing I did not do is revamp the "Time" parts of the Tf2 tutorials.
Those need to be rethought, but that will be a separate effort.

I retested all of these tutorials with these changes in place, and they
all seem to work for me.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit 5cd073459b71304e560e993dd74c462cf908c83f)

# Conflicts:
#	source/Tutorials/Intermediate/Tf2/Writing-A-Tf2-Static-Broadcaster-Cpp.rst
#	source/Tutorials/Intermediate/Tf2/Writing-A-Tf2-Static-Broadcaster-Py.rst