Fix sitemap generation.
It turns out that the sphinx sitemap plugin has a configuration
variable controlling how it creates the final URL to put
into the sitemap.  We always used the default in the past,
but for some reason this seems to have changed and no
longer works with our scheme.  So add in a custom sitemap
scheme, which matches what we actually do.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>