Fix dead links in the Iron release notes. (#3479)
That is, these demos are not available in Humble or
Foxy, so when we build the sphinx site with them, we
point at the wrong place.  Specifically point them
at the 'iron' branch, which fixes this problem.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
(cherry picked from commit c8db8341786a29f26468aab6c2e339e8e17a2fd9)