Merge the Fedora build instructions into the RHEL ones. (#3487)
The original Fedora installation page was written back in
the mists of time, and haven't changed a whole lot since
then.  However, the tooling to enable development of ROS 2
on Fedora *has* changed considerably (thanks in no small
part to the efforts of @cottsay).

In particular, developing on Fedora is 90% the same as
developing on RHEL, with the only difference being that
no additional repositories need to be enabled on Fedora.
So merge together the Fedora and RHEL installation
instructions.  Note that we also redirect the old
Fedora links to the RHEL installation page.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>