Add iceoryx_binding_c and rmw_connextdds to binary skip-keys.
On arm64, rmw_connextdds doesn't exist so that one makes sense
to skip.  For amd64, it does exist but it doesn't matter if
we skip the key; the binary archive is providing it.

I'm not quite sure why iceoryx_binding_c doesn't work as a
rosdep key, but like above it doesn't matter; the binary
archive is providing it.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>