Pin the versions of the dependencies. (#2725)
This installs as many dependencies as possible via the system
packages.  We then take the versions of those dependencies
and pin them in requirements.txt/constraints.txt .  This is
so that we keep a consistent set of dependencies that we know
will work for the lifetime of Ubuntu Jammy, without constantly
monitoring the upstream projects.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>
(cherry picked from commit 6461f099e363382dff37d7cc7d3dc689c4142cfc)