Sync fork (#17)
* Update launch files to declare WebotsController node (#11) (#3763)

* change launch files

* Update Setting-Up-Simulation-Webots-Basic.rst

* update minimal version tutorial 2

* add upcoming versions

* New Reset Handler tutorial (#12) (#3764)

* new tutorial

Co-authored-by: Olivier Michel <Olivier.Michel@cyberbotics.com>

* Update macOS-Development-Setup.rst (#3758)

- add Xcode App Store link
    - BASH -> ZSH (default)
    - make homebrew paths generic

Co-authored-by: Philipp Dittmann <dittmann@cs.uni-bremen.de>

* Change confusing title between Debian package and Debian platform (#3771)

Currently Debian does not offer ROS2 packages. Change the title that can
be confusing.

* Make the calendar more mobile-friendly. (#3774)

* Make the calendar more mobile-friendly.

When looking at some of the analytics, it looks like this
page is particularly not mobile friendly.  That's because
the Google calendar embedded on the bottom of the page
is of fixed with 800x600, which doesn't work well on
many mobile devices.

What we do here is to see if a device has 767 or less pixels.
If it does, then we just show the agenda, which fits in just
280 pixels.  If we have more than that, we have room for
the full monthly agenda, so we show the full thing.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
Co-authored-by: Tomoya Fujita <Tomoya.Fujita@sony.com>

---------

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
Co-authored-by: Olivier Michel <Olivier.Michel@cyberbotics.com>
Co-authored-by: PhDittmann <dittmann@tzi.de>
Co-authored-by: Philipp Dittmann <dittmann@cs.uni-bremen.de>
Co-authored-by: Quentin Quadrat <lecrapouille@gmail.com>
Co-authored-by: Chris Lalancette <clalancette@gmail.com>
Co-authored-by: Tomoya Fujita <Tomoya.Fujita@sony.com>