Cleanup the tf2 tutorials (#4056)
* Cleanup Introduction to Tf2.

In particular, make it less Linux centric.  Also cleanup
a few typos I noticed while reading.  And finally, change
the packages we install to be the ones actually needed
by the tutorial nowadays.

* Cleanup Writing-A-Tf2-Broadcaster-* tutorials.

Just make the C++ and the Python one a little more similar,
and fix a few minor typos.

* Minor cleanup in Writing-A-Tf2-Static-Broadcaster-* tutorials.

* Small cleanups in the Writing-A-Tf2-Listener tutorials.

Make the Python and C++ tutorials more alike.

* Minor updates to the Adding-A-Frame tutorials.

Just to make the Python and C++ tutorials more similar.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>