Update Colcon-Tutorial.rst
Added instructions for how to run the tutorials if you are on the core version of ROS2.  Updated because there will be people who install a minimal version of ROS2 who want to follow the tutorials. This could include users following the [ROS2 Docker instructions](https://docs.ros.org/en/rolling/How-To-Guides/Setup-ROS-2-with-VSCode-and-Docker-Container.html).

Signed-off-by: BrnBlrg <BenjaminAaronBlumer@gmail.com>