Remove duplicate argument in call to `ros2 pkg create` (#4103)
The tutorial page to create the first ROS2 package instructs users to
pass the `--license` argument twice when calling `ros2 pkg create`.

Using `--license` twice has the effect that the first license in the
argument list is ignored. There's no point in adding this twice, it just
confuses users and makes the line longer.

This commit removes the second `--license Apache-2.0` in the line for
both the CMake and Python variants of the command.