Improving possible confusing instructions (#4124)
* Improving possible confusing instructions

The instruction tell the reader to search and replace USERNAME in 2 scripts. But some USERNAME strings should not be replaced.
Changed the string that need to be replaced with YOUR_USERNAME and changed the instructions to match that.

Signed-off-by: Tim <timbosschaerts@gmail.com>
Co-authored-by: Tomoya Fujita <Tomoya.Fujita@sony.com>
(cherry picked from commit fecf496f726ef269c45b681421834d38e2d68f9f)