Rename "Package maintainer guide" to "Core maintainer guide".
This rename is motivated by a desire to begin a manual / guide for
maintainers of _community_ packages in ROS 2 distributions.

The current title is overly broad, as the guide content is targeted at
ROS 2 core package maintainers rather than a maintainer of any package
released in ROS 2.

Renaming and redirecting articles has a cost and I know this rename
isn't free. But I think it's justified in order to provide space for
content aimed at community maintainers.