Update Ada client library status for jazzy
My apologies if this is not the proper venue, or there are specific requirements for a client library to be considered maintained.

I've just updated the Ada library to work on `jazzy`, and I'm generally available to attend issues about this library (which I authored). 

A related question is: should this PR be (also/instead) opened against the `jazzy` branch?

Signed-off-by: Alejandro R Mosteo <amosteo@unizar.es>