add slightly more exposition in the Overview section
Sure, this more or less expected for the reader to already know.

It's also some slight duplication, since it's mentioned further down
that introspection is disabled by default. I figure it's better to say
it at the very start though.

This way I feel the "why would I need service introspection?" is
addressed as well.

Signed-off-by: Daniel Cranston <cranston.daniel@gmail.com>