lift 'remote node' one level higher, add event callback explanation
It currently exists as a subsection of "3 Build and run", which makes it
somewhat hidden. It also makes it hidden from the ToC.

This commit lifts it one level higher, as a subsection to a new
"Extensions" section.

Further, a second subsection to "Extensions" is added that explains that
event callbacks exists, and how to use them.