Update Windows-Development-Setup.rst with note about very long build times
I tested the source installation on Windows 10 machine with ROS2 Kilted as part of https://github.com/ros2/kilted_tutorial_party/issues/297 where the `colcon build --merge-install` took forever that I had to leave the installation overnight. It might be good to add a brief note about this in the installation instructions.

Signed-off-by: Mohammed Tousif Zaman <zamantousif@gmail.com>